+++ src/lib/private/layers/GeoJSONLayer.js
