+++ src/lib/private/widgets/BasemapGallery.js
